[Train] Split test_gpu tests (#29297) (#29410)
We had AMI changes that happened around same time test_gpu times out but for our ongoing work we should keep updated with latest pytorch and cuda 11.6.

Therefore this change split tests into two files so single buck target won't timeout.
